TN7-272 [FEAT] 설문 분석 페이지 구현(선형배율, 감정바)
