+++ src/components/common/Frame.js
@@ -28,7 +28,7 @@ const MainWrapper = styled.div`
   margin: 0 auto;
 `
 
-const Frame = () => {
+const Frame = () => { //     
   return (
     <>
       <MainWrapper>
